+++ src/Commands/Syntex/UnPublishSyntexModel.cs
@@ -2,7 +2,8 @@
 using PnP.Framework;
 using PnP.PowerShell.Commands.Base;
 using PnP.PowerShell.Commands.Base.PipeBinds;
-using System;
+using PnP.PowerShell.Commands.Model;
+using PnP.PowerShell.Commands.Model.Syntex;
 using System.Management.Automation;
 
 namespace PnP.PowerShell.Commands.Syntex
@@ -10,14 +11,30 @@ namespace PnP.PowerShell.Commands.Syntex
     [Cmdlet(VerbsData.Unpublish, "PnPSyntexModel")]
     public class UnPublishSyntexModel : PnPWebCmdlet
     {
+        const string ParameterSet_SINGLE = "Single";
+        const string Parameterset_BATCHED = "Batched";
+
         [Parameter(Mandatory = true, ValueFromPipeline = true, Position = 0)]
         public SyntexModelPipeBind Model;
 
-        [Parameter(Mandatory = true)]
+        [Parameter(Mandatory = true, ParameterSetName = ParameterSet_SINGLE)]
         public string ListWebUrl;
 
-        [Parameter(Mandatory = true)]
+        [Parameter(Mandatory = true, ParameterSetName = ParameterSet_SINGLE)]
         public ListPipeBind List;
+
+        [Parameter(Mandatory = true, ParameterSetName = Parameterset_BATCHED)]
+        public string TargetSiteUrl;
+
+        [Parameter(Mandatory = true, ParameterSetName = Parameterset_BATCHED)]
+        public string TargetWebServerRelativeUrl;
+
+        [Parameter(Mandatory = true, ParameterSetName = Parameterset_BATCHED)]
+        public string TargetLibraryServerRelativeUrl;
+
+        [Parameter(Mandatory = true, ParameterSetName = Parameterset_BATCHED)]
+        [ValidateNotNull]
+        public PnPBatch Batch;
 
         protected override void ExecuteCmdlet()
         {
@@ -25,29 +42,64 @@ namespace PnP.PowerShell.Commands.Syntex
 
             if (ctx.Web.IsSyntexContentCenter())
             {
-                // Get the model we're publishing
-                ISyntexModel modelToUnPublish = Model.GetSyntexModel();
-
-                if (modelToUnPublish == null)
+                if (ParameterSpecified(nameof(Batch)))
                 {
-                    throw new ArgumentException("Provide a valid model to unpublish");
-                }
+                    // Get the model we're publishing
+                    ISyntexModel modelToPublish = Model.GetSyntexModel(Batch);
 
-                // resolve the list 
-                IList listToUnPublishModelFrom = null;
-                using (var listContext = PnPConnection.Current.CloneContext(ListWebUrl))
+                    if (modelToPublish == null)
+                    {
+                        throw new PSArgumentException("Provide a valid model to publish");
+                    }
+
+                    modelToPublish.UnPublishModelBatch(Batch.Batch, new SyntexModelUnPublishOptions()
+                    {
+                        TargetSiteUrl = TargetSiteUrl,
+                        TargetWebServerRelativeUrl = TargetWebServerRelativeUrl,
+                        TargetLibraryServerRelativeUrl = TargetLibraryServerRelativeUrl,
+                    });
+                }
+                else
                 {
-                    var pnpContext = PnPCoreSdk.Instance.GetPnPContext(listContext);
-                    listToUnPublishModelFrom = List.GetList(pnpContext);
+                    // Get the model we're publishing
+                    ISyntexModel modelToUnPublish = Model.GetSyntexModel();
+
+                    if (modelToUnPublish == null)
+                    {
+                        throw new PSArgumentException("Provide a valid model to unpublish");
+                    }
+
+                    // resolve the list 
+                    IList listToUnPublishModelFrom = null;
+                    using (var listContext = PnPConnection.Current.CloneContext(ListWebUrl))
+                    {
+                        var pnpContext = PnPCoreSdk.Instance.GetPnPContext(listContext);
+                        listToUnPublishModelFrom = List.GetList(pnpContext);
+                    }
+
+                    if (listToUnPublishModelFrom == null)
+                    {
+                        throw new PSArgumentException("Provide a valid list to unpublish the Syntex model from");
+                    }
+
+                    var unPublishResult = modelToUnPublish.UnPublishModel(listToUnPublishModelFrom);
+                    if (unPublishResult != null)
+                    {
+                        WriteObject(new SyntexPublicationResult()
+                        {
+                            ErrorMessage = unPublishResult.ErrorMessage,
+                            StatusCode = unPublishResult.StatusCode,
+                            Publication = new Model.Syntex.SyntexModelPublication()
+                            {
+                                ModelUniqueId = unPublishResult.Publication.ModelUniqueId,
+                                TargetSiteUrl = unPublishResult.Publication.TargetSiteUrl,
+                                TargetWebServerRelativeUrl = unPublishResult.Publication.TargetWebServerRelativeUrl,
+                                TargetLibraryServerRelativeUrl = unPublishResult.Publication.TargetLibraryServerRelativeUrl,
+                                ViewOption = unPublishResult.Publication.ViewOption
+                            }
+                        });
+                    }
                 }
-
-                if (listToUnPublishModelFrom == null)
-                {
-                    throw new ArgumentException("Provide a valid list to unpublish the Syntex model from");
-                }
-
-                WriteObject(modelToUnPublish.UnPublishModel(listToUnPublishModelFrom));                
-
             }
             else
             {