+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Added `Request-PnPSyntexClassifyAndExtract` cmdlet to request classification and extraction of a file or all files in a list
 - Added `Get-PnPSyntexModel` cmdlet to list the defined SharePoint Syntex models in a SharePoint Syntex content center site
 - Added `Publish-PnPSyntexModel` cmdlet to publish a SharePoint Syntex model to a library
 - Added `Unpublish-PnPSyntexModel` cmdlet to unpublish a SharePoint Syntex model from a library
@@ -17,7 +18,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 ### Contributors
-
+- Bert Jansen [jansenbe]
+  
 ## [1.4.0]
 
 ### Added